"Fix" retroactive_revinfo in LTV validator
The fix is a kludge, and it's becoming more and more clear that the
current design of the LTV validation internals _and_ the API aren't
going to last long. Consider deprecating it in the next release if
we have an ETSI EN 319 102-1 compliant implementation to replace it,
which can also be put in some kind of legacy mode where it'd use the
same trust model as the current LTV validation logic.